refactor(collectors): pull up shouldIgnoreTransaction and getTransaction methods to parent from child classes to avoid duplication
