+++ src/Collectors/CommandCollector.php
@@ -8,7 +8,6 @@ use Illuminate\Console\Events\CommandFinished;
 use Illuminate\Console\Events\CommandStarting;
 use Illuminate\Support\Facades\Log;
 use PhilKra\Events\Transaction;
-use PhilKra\Exception\Transaction\UnknownTransactionException;
 use Throwable;
 
 /**
@@ -50,15 +49,6 @@ class CommandCollector extends EventDataCollector implements DataCollector
                 }
             }
         });
-    }
-
-    protected function getTransaction(string $transaction_name): ?Transaction
-    {
-        try {
-            return $this->agent->getTransaction($transaction_name);
-        } catch (UnknownTransactionException $e) {
-            return null;
-        }
     }
 
     protected function startTransaction(string $transaction_name): Transaction
@@ -115,12 +105,5 @@ class CommandCollector extends EventDataCollector implements DataCollector
         }
 
         return $this->shouldIgnoreTransaction($transaction_name) ? '' : $transaction_name;
-    }
-
-    protected function shouldIgnoreTransaction(string $transaction_name): bool
-    {
-        $pattern = $this->config->get('elastic-apm-laravel.transactions.ignorePatterns');
-
-        return $pattern && preg_match($pattern, $transaction_name);
     }
 }
